Reimplement skipped initialize tests
